pyproject.toml
[tool.poetry]
name = "freshli-agent-python"
version = "0.1.0"
description = "The Python language agent used by the Freshli CLI"
license = "AGPL-3.0-or-later"
authors = ["M. Scott Ford <scott@corgibytes.com>"]
readme = "README.md"
repository = "https://github.com/corgibytes/freshli-agent-python"
packages = [{include = "corgibytes", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
typer = "^0.9.0"
grpcio = "^1.54.0"
grpcio-tools = "^1.54.0"
grpcio-health-checking = "^1.54.2"
grpcio-reflection = "^1.54.2"

[tool.poetry.scripts]
freshli-agent-python = "corgibytes.freshli.agent:app"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-expecter = "^3.0"
pytest-cov = "^4.1.0"
pytest-datadir = "^1.4.1"
pytest-describe = "^2.1.0"
pyinstaller = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
